Deprecated w-on* attributes
Added support for on*() attributes.
