+++ compiler/Parser.js
@@ -1,6 +1,6 @@
 'use strict';
 var ok = require('assert').ok;
-var AttributePlaceholder = require('./ast/AttributePlaceholder');
+var replacePlaceholderEscapeFuncs = require('./util/replacePlaceholderEscapeFuncs');
 
 var COMPILER_ATTRIBUTE_HANDLERS = {
     'preserve-whitespace': function(attr, context) {
@@ -15,24 +15,6 @@ var ieConditionalCommentRegExp = /^\[if [^]*?<!\[endif\]$/;
 
 function isIEConditionalComment(comment) {
     return ieConditionalCommentRegExp.test(comment);
-}
-
-function replacePlaceholderEscapeFuncs(node, context) {
-    var walker = context.createWalker({
-        exit: function(node, parent) {
-            if (node.type === 'FunctionCall' &&
-                node.callee.type === 'Identifier') {
-
-                if (node.callee.name === '$noEscapeXml') {
-                    return new AttributePlaceholder({escape: false, value: node.args[0]});
-                } else if (node.callee.name === '$escapeXml') {
-                    return new AttributePlaceholder({escape: true, value: node.args[0]});
-                }
-            }
-        }
-    });
-
-    return walker.walk(node);
 }
 
 function mergeShorthandClassNames(el, shorthandClassNames, context) {
