+++ test/autotests/taglib-lookup/forEachAttribute/expected.json
@@ -28,5 +28,6 @@
     "no-update-if",
     "no-update-body-if",
     "w-preserve-attrs",
+    "on*",
     "w-on*"
 ]
