+++ test/autotests/widgets-browser/update-manager-batch-updates/template.marko
@@ -1,4 +1,4 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind
     include()></button>
