+++ test/autotests/widgets-browser/widget-destroy-unsubscribe-dom-events/template.marko
@@ -1,6 +1,6 @@
 <div w-bind>
     <a href="#foo" ref="fooLink"
-        w-ondblclick="handleFooLinkDblClick"
-        w-onmouseout="handleFooLinkMouseOut">
+        onDblclick("handleFooLinkDblClick")
+        onMouseout("handleFooLinkMouseOut")>
     </a>
 </div>
