+++ test/autotests/widgets-browser/widget-dom-event-listeners-repeated-non-bubbling/template.marko
@@ -1,6 +1,6 @@
 <div w-bind>
 	<ul>
-		<li w-onMouseMove="handleMouseMove" ref="foo-${loop.getIndex()}"
+		<li onMouseMove("handleMouseMove") ref="foo-${loop.getIndex()}"
 			 for(color in ['red', 'green', 'blue'] | status-var=loop)>${color}</li>
 	</ul>
 </div>
