+++ test/autotests/widgets-browser/widget-dom-events/components/app-legacy-button/template.marko
@@ -1,5 +1,5 @@
 <button type="button" class="app-legacy-button"
     w-bind="./widget"
-    w-onmousedown="handleRootMouseDown">
+    onMousedown("handleRootMouseDown")>
     <invoke data.renderBody(out)/>
 </button>
