+++ test/autotests/widgets-browser/widget-event-handler-method-conditional-bubbles/template.marko
@@ -1,5 +1,5 @@
 <div w-bind>
-	<input w-onclick=(false && 'handleButtonClick') ref="inputWithoutHandler" type="button">
-	<input w-onclick=(true && 'handleButtonClick') ref="inputWithHandler" type="button">
-	<input w-onclick='handleButtonClick' ref="inputWithLiteralHandler" type="button">
+	<input onClick(false && 'handleButtonClick') ref="inputWithoutHandler" type="button">
+	<input onClick(true && 'handleButtonClick') ref="inputWithHandler" type="button">
+	<input onClick('handleButtonClick') ref="inputWithLiteralHandler" type="button">
 </div>
