+++ test/autotests/widgets-browser/widget-event-handler-method-conditional-direct/template.marko
@@ -1,5 +1,5 @@
 <div w-bind>
-	<input w-onMouseMove=(false && "handleMouseMove") ref="inputWithoutHandler" type="button">
-	<input w-onMouseMove=(true && 'handleMouseMove') ref="inputWithHandler" type="button">
-	<input w-onMouseMove='handleMouseMove' ref="inputWithLiteralHandler" type="button">
+	<input onMouseMove(false && "handleMouseMove") ref="inputWithoutHandler" type="button">
+	<input onMouseMove(true && 'handleMouseMove') ref="inputWithHandler" type="button">
+	<input onMouseMove('handleMouseMove') ref="inputWithLiteralHandler" type="button">
 </div>
