+++ test/autotests/widgets-browser/widget-event-handler-method-dynamic/template.marko
@@ -1,7 +1,7 @@
 <div w-bind>
 	<h1>app-macro-events</h1>
 	<macro renderButton(id, label, handler)>
-		<button ref="${id}" type="button" w-onClick="${handler}">
+		<button ref="${id}" type="button" onClick("${handler}")>
 			$label
 		</button>
     </macro>
