+++ test/autotests/widgets-browser/widget-extend/components/app-extend-button/template.marko
@@ -1,5 +1,5 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind="./widget">
 
     <if(data.label)>
