+++ test/autotests/widgets-browser/widget-rerender-reuse-stateful/components/app-stateful-button/template.marko
@@ -1,4 +1,4 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind
     include()></button>
