+++ test/autotests/widgets-browser/widget-stateful-copy-state-on-write/template.marko
@@ -1,4 +1,4 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind
     include></button>
