+++ test/autotests/widgets-browser/widget-stateful-no-copy-state-on-write-if-same-value/template.marko
@@ -1,4 +1,4 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind
     include()></button>
