+++ test/autotests/widgets-browser/widget-stateful-reuse-widgets/components/app-stateful-button/template.marko
@@ -1,4 +1,4 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind
     include()></button>
