+++ test/autotests/widgets-browser/widget-stopPropagation/template.marko
@@ -1,7 +1,7 @@
 <div w-bind>
-	<div w-onclick="handleDivClick">
+	<div onClick("handleDivClick")>
 		Hello
-		<button ref="button" type="button" w-onclick="handleButtonClick">
+		<button ref="button" type="button" onClick("handleButtonClick")>
 			Button
 		</button>
 	</div>
