+++ test/autotests/widgets-pages/dom-events/components/app-foo/template.marko
@@ -1,2 +1,2 @@
 div w-bind
-    button ref="button" w-onMouseMove='handleButtonMouseMove' w-onClick="handleButtonClick"
+    button ref="button" onMouseMove('handleButtonMouseMove') onClick("handleButtonClick")
