+++ test/autotests/widgets-pages/split-widget-renderer/components/app-button-split/template.marko
@@ -1,4 +1,4 @@
 <button ${data.rootAttrs}
-    w-onclick="handleClick"
+    onClick("handleClick")
     w-bind
     w-body></button>
